# compact.Tree: Change semantic of adding leaves