--- a/merkle/compact/tree_test.go
+++ b/merkle/compact/tree_test.go
@@ -83,8 +83,8 @@ func TestAddingLeaves(t *testing.T) {
 			idx := 0
 			for _, br := range tc.breaks {
 				for ; idx < br; idx++ {
-					if _, err := tree.AddLeaf(inputs[idx], nil); err != nil {
-						t.Fatalf("AddLeaf: %v", err)
+					if _, err := tree.AppendLeaf(inputs[idx], nil); err != nil {
+						t.Fatalf("AppendLeaf: %v", err)
 					}
 					if err := checkUnusedNodesInvariant(tree); err != nil {
 						t.Fatalf("UnusedNodesInvariant check failed: %v", err)
@@ -195,7 +195,7 @@ func TestCompactVsFullTree(t *testing.T) {
 		newLeaf := []byte(fmt.Sprintf("Leaf %d", i))
 
 		iSeq, iHash := imt.AddLeaf(newLeaf)
-		cHash, err := cmt.AddLeaf(newLeaf, func(id NodeID, hash []byte) {
+		cHash, err := cmt.AppendLeaf(newLeaf, func(id NodeID, hash []byte) {
 			nodes[id] = hash
 		})
 		if err != nil {
@@ -222,9 +222,9 @@ func TestCompactVsFullTree(t *testing.T) {
 	cmt := NewTree(rfc6962.DefaultHasher)
 	for i := int64(0); i < imt.LeafCount(); i++ {
 		newLeaf := []byte(fmt.Sprintf("Leaf %d", i))
-		_, err := cmt.AddLeaf(newLeaf, nil)
+		_, err := cmt.AppendLeaf(newLeaf, nil)
 		if err != nil {
-			t.Fatalf("AddLeaf(%d)=_,_,%v, want _,_,nil", i, err)
+			t.Fatalf("AppendLeaf(%d)=_,_,%v, want _,_,nil", i, err)
 		}
 		if got, want := cmt.Size(), i+1; got != want {
 			t.Fatalf("new tree size=%d, want %d", got, want)
@@ -260,7 +260,7 @@ func TestRootHashForVariousTreeSizes(t *testing.T) {
 		tree := NewTree(rfc6962.DefaultHasher)
 		for i := int64(0); i < test.size; i++ {
 			l := []byte{byte(i & 0xff), byte((i >> 8) & 0xff)}
-			tree.AddLeaf(l, nil)
+			tree.AppendLeaf(l, nil)
 		}
 		if gotRoot := mustGetRoot(t, tree); !bytes.Equal(gotRoot, test.wantRoot) {
 			t.Errorf("Test (treesize=%v) got root %v, want %v", test.size, b64e(gotRoot), b64e(test.wantRoot))
@@ -283,23 +283,23 @@ func TestRootHashForVariousTreeSizes(t *testing.T) {
 	}
 }
 
-func benchmarkAddLeafHash(b *testing.B, visit VisitFn) {
+func benchmarkAppendLeaf(b *testing.B, visit VisitFn) {
 	const size = 1024
 	for n := 0; n < b.N; n++ {
 		tree := NewTree(rfc6962.DefaultHasher)
 		for i := 0; i < size; i++ {
 			l := []byte{byte(i & 0xff), byte((i >> 8) & 0xff)}
-			if _, err := tree.AddLeaf(l, visit); err != nil {
-				b.Fatalf("AddLeaf: %v", err)
+			if _, err := tree.AppendLeaf(l, visit); err != nil {
+				b.Fatalf("AppendLeaf: %v", err)
 			}
 		}
 	}
 }
 
-func BenchmarkAddLeafHash(b *testing.B) {
-	benchmarkAddLeafHash(b, func(NodeID, []byte) {})
+func BenchmarkAppendLeaf(b *testing.B) {
+	benchmarkAppendLeaf(b, func(NodeID, []byte) {})
 }
 
-func BenchmarkAddLeafHashNoVisitor(b *testing.B) {
-	benchmarkAddLeafHash(b, nil)
+func BenchmarkAppendLeafNoVisitor(b *testing.B) {
+	benchmarkAppendLeaf(b, nil)
 }