This change introduces a family of `AppendLeaf` methods in `compact.Tree`,
and removes the old `AddLeaf` methods.

The new `AppendLeaf` methods are equivalent to `AddLeaf` methods, but they
do not report ephemeral nodes along the right border of the Merkle tree.
We don't simply reuse the old method names because the semantic is changed.

Benefits of `AppendLeaf` over `AddLeaf` methods are:
- They are more aligned with `compact.Range` semantic.
- If called multiple times, `AddLeaf` wastes CPU on computing ephemeral
  hashes that will be overridden by subsequent calls. Instead, ephemeral
  nodes can be computed lazily in the end, after all `AppendLeaf`-s. The
  time complexity of doing N appends reduces from `O(N log N)` to `O(N)`.

### Benchmarks
#### Before:
```
BenchmarkAddLeafHash-12             	     200	   9189829 ns/op
BenchmarkAddLeafHashNoVisitor-12    	     200	   9259209 ns/op
```
#### After:
```
BenchmarkAppendLeaf-12             	     500	   2770825 ns/op
BenchmarkAppendLeafNoVisitor-12    	     500	   2967951 ns/op
```

<!---
Describe your changes in detail here.
If this fixes an issue, please write "Fixes #123", substituting the issue number.
-->

### Checklist

<!---
Go over all the following points, and put an `x` in all the boxes that apply.
Feel free to not tick any boxes that don't apply to this PR (e.g. refactoring may not need a CHANGELOG update).
If you're unsure about any of these, don't hesitate to ask. We're here to help!
-->

- [x] I have updated the [CHANGELOG](CHANGELOG.md).
  - Adjust the draft version number according to [semantic versioning](https://semver.org/) rules.
- [ ] I have updated [documentation](docs/) accordingly.