@AlCutter 2019-05-13
Looks good, but one thing which is a little unnerving is that I could naively do something like this and really hurt myself:

1. add some leaves (--> leaf hashes get set)
1. call CurrentRoot() (--> intermediates get set)
1. add some more leaves (--> more leaf hashes get set)
1. not call CurrentRoot() (--> uh oh...)

i.e I _must_ call CurrentRoot if I call AppendLeaf (and again if I call AppendLeaf after having called CurrentRoot) -- or I _must_ abort the TX.