@Martin2112 2019-05-16
Before I do the review this package is generally useful so could we document the add / root contract a bit more in the godoc? For example the doc implies `CalculateRoot` is not idempotent, would help users if it said so directly.