@pavelkalinnikov 2019-05-16
`CalculateRoot` is idempotent with respect to `compact.Tree` itself (it merely hashes the hashes). But if the `visit` callback is not idempotent then `CalculateRoot(visit)` is likewise not (with respect to visitor's side effects) - which is a common sense. I'll see how I can improve the commenting, but please don't be blocked on it and also feel free to suggest if you have concrete ideas.