@Martin2112 2019-05-16
It doesn't need pages of text, just a few clarifications or an example function maybe. Think about about it from the point of view of someone who wants to use this as a library and doesn't know our conventions. Definitely no need to go into visitor side effects.