@pavelkalinnikov 2019-05-16
The coverage thing is probably due to having the number of lines differ before and after the change, even though logically nothing changed.